# apps/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Adversary backend API — scenarios, geo, runs (WebSocket ops+map), manage, generate/route, AUTH_MODE=off|session. Live channels: GET /v1/runs/{id}/ws/ops and /ws/map (Upgrade). Prometheus scrape: GET /metrics (not listed as OpenAPI paths). Architecture: docs/api.md.",
        "title": "Adversary API",
        "contact": {},
        "version": "0.9.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/manage/scenarios": {
            "get": {
                "description": "Paginated list with payload size_bytes, target/event counts, owner, active run count. AUTH_MODE=off: ownerUserId usually null.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manage"
                ],
                "summary": "Manage: list scenarios (storage)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "draft or ready",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Name search (ILIKE)",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 50, max 200)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_shammalie_adversary_apps_api_internal_store.ManageListResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    }
                }
            }
        },
        "/v1/manage/scenarios/bulk-delete": {
            "post": {
                "description": "Same cascade-stop semantics as single delete. Partial success returns 200 with failed map.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manage"
                ],
                "summary": "Manage: bulk delete scenarios",
                "parameters": [
                    {
                        "description": "Scenario ids",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.bulkDeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.bulkDeleteResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    }
                }
            }
        },
        "/v1/manage/scenarios/{id}": {
            "delete": {
                "description": "Cascade-stops active runs for the scenario, deletes all runs referencing it, then deletes the scenario.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manage"
                ],
                "summary": "Manage: delete scenario",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Scenario ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.manageDeleteResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    }
                }
            }
        },
        "/v1/manage/stats": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manage"
                ],
                "summary": "Manage: storage / run stats",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_shammalie_adversary_apps_api_internal_store.ManageStats"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_shammalie_adversary_apps_api_internal_store.ManageListResult": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_shammalie_adversary_apps_api_internal_store.ManageScenarioRow"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "github_com_shammalie_adversary_apps_api_internal_store.ManageScenarioRow": {
            "type": "object",
            "properties": {
                "activeRuns": {
                    "type": "integer"
                },
                "eventCount": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "ownerUserId": {
                    "type": "string"
                },
                "sizeBytes": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "targetCount": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_shammalie_adversary_apps_api_internal_store.ManageStats": {
            "type": "object",
            "properties": {
                "draftCount": {
                    "type": "integer"
                },
                "readyCount": {
                    "type": "integer"
                },
                "runsActive": {
                    "type": "integer"
                },
                "runsCompleted": {
                    "type": "integer"
                },
                "runsStopped": {
                    "type": "integer"
                },
                "scenarioCount": {
                    "type": "integer"
                },
                "totalPayloadBytes": {
                    "type": "integer"
                }
            }
        },
...
        "internal_handler.bulkDeleteRequest": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "internal_handler.bulkDeleteResult": {
            "type": "object",
            "properties": {
                "deleted": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "failed": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "internal_handler.errorBody": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler.manageDeleteResponse": {
            "type": "object",
            "properties": {
                "deletedRuns": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "stoppedRuns": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "type": "apiKey",
            "name": "adversary_session",
            "in": "cookie"
        }
    }
}